procedural animations learning
i tried a couple o' things with a simple rigged character and came to a conclusion that I WILL NOT try to animate a detailed character no matter what as it is a tedious and hard task. So for the time being I am sticking to a simple mesh and rig.
Yeah I also deleted some useless material